dom/tv/test/mochitest/test_tv_get_channels_during_scanning.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test an error case for GetChannels during scanning for TV API</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript" src="./head.js"></script>
<script type="application/javascript">

function runTest() {
  ok('tv' in navigator, "navigator.tv should exist.");

  navigator.tv.getTuners().then(
  function(aTuners) {
    ok(aTuners.length > 0, "Got at least 1 tuner.");

    aTuners[0].getSources().then(
      function(aSources) {
        ok(aSources.length > 0, "Got at least 1 source.");
        var source = aSources[0];

        // TODO Bug 1088818 - Modify the behavior of channel scanning.
        source.startScanning({}).then(
          function() {
            source.getChannels().then(
              function() {
                ok(false, "Getting channels during scanning should get error.");
                  SimpleTest.finish();
                },
                function(aError) {
                  is(aError.name, "InvalidStateError",
                     "InvalidStateError should be expected.");
                  SimpleTest.finish();
                }
              );
            },
            function(aError) {
              ok(false, "Error occurred when starting scanning: " + aError);
              SimpleTest.finish();
            }
          );
        },
        function(aError) {
          ok(false, "Error occurred when getting sources: " + aError);
          SimpleTest.finish();
        }
      );
    },
    function(aError) {
      ok(false, "Error occurred when getting tuners: " + aError);
      SimpleTest.finish();
    }
  );
}

SimpleTest.waitForExplicitFinish();
prepareTest(runTest);

</script>
</pre>
</body>
</html>
